pyproject.toml
[project]
name = "imagefeatures"
version = "0.1.0"
description = "Classic image feature extraction for Python"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Michael"}
]
keywords = ["image", "features", "cbir", "texture", "color", "histogram"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "numpy>=1.20",
    "Pillow>=8.0",
]

[project.optional-dependencies]
cli = ["click>=8.0"]
dev = ["pytest>=7.0"]

[project.scripts]
imagefeatures = "imagefeatures.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["imagefeatures*"]
